+++ app/src/main/java/com/zeapo/pwdstore/utils/Result.kt
@@ -13,4 +13,6 @@ sealed class Result {
 
     object Ok : Result()
     data class Err(val err: Exception) : Result()
+
+    fun isOk() = this is Ok
 }
